requirements_streamlit.txt
# Cassava Leaf Disease Classification - Streamlit Requirements
# Run: pip install -r requirements_streamlit.txt

streamlit>=1.22.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.5.0
tensorflow>=2.9.0
Pillow>=9.0.0
